move login template from components to templates
